documents/requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
Werkzeug==2.0.3
SQLAlchemy==1.4.32
flask-restx
PyJWT